Web-Interface/requirements.txt
# Web Interface Dependencies - Python 3.13.5 Compatible
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy>=2.0.36
aiosqlite>=0.20.0
pydantic>=2.9.0
python-multipart>=0.0.12
pyyaml>=6.0.2
psutil>=6.0.0
paho-mqtt>=2.0.0
